Change int to size_t

+++ src/log/log_harness.cpp
@@ -16,7 +16,7 @@ LogHarness::~LogHarness()
   google::ShutdownGoogleLogging();
 }
 
-void LogHarness::handleSingal(const char* data, int size)
+void LogHarness::handleSingal(const char* data, size_t size)
 {
   LOG(FATAL) << std::string(data, size);
 }
